updated export python scripts

+++ export/export_data.sh
@@ -17,17 +17,21 @@ if [ ! -d "$OUTDIR" ]; then
     mkdir $OUTDIR
 fi
 
-# Remove any existing files in output directory
-rm $OUTDIR/*
+# If OUTDIR is not empty
+if [ "$(ls -A $OUTDIR)" ]; then
+
+   # Remove existing files
+   rm $OUTDIR/*
+fi
 
 # Export user data to $OUTDIR/users.csv using python script
 ./export_users.py $DATABASE $OUTDIR
 
 # Export facilities data to $OUTDIR/facilities.csv using python script
-./export_facilities.py $DATABASE $OUTDIR
+# ./export_facilities.py $DATABASE $OUTDIR
 
 # Export assets data to $OUTDIR/assets.csv using python script
-./export_assets.py $DATABASE $OUTDIR
-
+# ./export_assets.py $DATABASE $OUTDIR
+ 
 # Export transfers data to $OUTDIR/transfer.csv using python script
-./export_transfers.py $DATABASE $OUTDIR
+# ./export_transfers.py $DATABASE $OUTDIR
